Fix warnings: disable Spectre mitigation, fix INF, fix intellisense
For INF file fix see:
* https://github.com/dokan-dev/dokany/issues/648
* https://docs.microsoft.com/en-us/windows-hardware/drivers/install/inf-delfiles-directive
```
Do not specify a file that is listed in a CopyFiles directive.
If a file is listed in both a CopyFiles-referenced and a
DelFiles-referenced section, and the file is currently present on the
system with a valid signature, the operating system might optimize away
the copy operation but perform the delete operation.
This is very likely not what the INF writer intended.
```